Wake the parent when a worker dies inside its turn, not just via failSession
Bugbot (high severity): the parent completion wake on a failed worker
only ran from the failSession activity, but the common failure path ends
inside runAgentTurn — it marks the session failed and returns "failed",
after which the session workflow exits WITHOUT calling failSession or
markSessionIdle. So a worker that died inside its turn (model error,
non-retryable provider failure, max-turns terminal) never woke its
manager.

Fire the parent wake from runAgentTurn's non-retryable failed branch
too. Extract the wake helper (notifyParentOfChildTerminal + text/payload
builders) into a dedicated parent-wake module so both the session-state
terminal activities and runAgentTurn can call it without a circular
import between those two activity modules. The wake stays idempotent per
terminal episode on the child's lastSequence, so the in-turn wake and any
workflow-level wake for the same failure never double-fire.

Adds an integration test: a worker scripted to throw mid-turn reaches
"failed" via runAgentTurn (failSession never runs) and the manager is
still woken exactly once with a FAILED wake.
